requirements.txt
boto==2.48.0
boto3==1.7.14
botocore==1.10.14
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
docutils==0.14
en-core-web-lg==2.0.0
en-core-web-sm==2.0.0
gensim==3.4.0
idna==2.6
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.2
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
psutil==5.4.5
pycorenlp==0.3.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
smart-open==1.5.7
spacy==2.0.11
stanfordcorenlp==3.9.1.1
termcolor==1.1.0
textblob==0.15.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.0
ujson==1.35
urllib3==1.22
wrapt==1.10.11
xgboost==0.71
